Restructure diagnostics
